Prepare for common node port pool
A node must itself manage which nodes and ports it is connected to and look
up itself those on every `on_process`.

I think this will help decomplicate the supervisor that executes the nodes,
and allows a node to peek into other nodes, maybe even dynamically without
actually connected to it (e.g when user makes measurements when synth is
running).